Move structs which are send across threads to own module
